requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio
scikit-learn
optimum
numpy<2
tuna
pystray
Pillow
faiss-cpu
einops
sentence-transformers
transformers
pywin32
dotenv